requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
Django==1.11
django-datatables-view==1.14.0
djangorestframework==3.6.3
idna==2.6
psycopg2==2.7.3
pytz==2017.2
requests==2.18.4
urllib3==1.22
